requirements.txt
torch
transformers
biopython
rich
